+++ src/main/java/org/opentripplanner/updater/vehicle_position/VehiclePositionsUpdaterParameters.java
@@ -1,7 +1,7 @@
 package org.opentripplanner.updater.vehicle_position;
 
 import java.net.URI;
-import java.util.Map;
+import java.time.Duration;
 import java.util.Objects;
 import org.opentripplanner.updater.spi.HttpHeaders;
 import org.opentripplanner.updater.spi.PollingGraphUpdaterParameters;
@@ -10,7 +10,7 @@ public record VehiclePositionsUpdaterParameters(
   String configRef,
   String feedId,
   URI url,
-  int frequencySec,
+  Duration frequency,
   HttpHeaders headers
 )
   implements PollingGraphUpdaterParameters {